data-fetcher/requirements.txt
requests==2.31.0
python-dateutil==2.8.2
tqdm==4.66.1